Showed count with pagination and added eager loading

+++ app/controllers/api/v1/artists_collaborators_controller.rb
@@ -27,7 +27,7 @@ class Api::V1::ArtistsCollaboratorsController < Api::BaseController
   param_group :doc_update_access
   def update_access
     if @artists_collaborator.update(access: params[:access])
-      render json: current_user.collaborators, each_serializer: Api::V1::CollaboratorSerializer
+      render json: current_user.collaborators_details.includes(:collaborator), each_serializer: Api::V1::CollaboratorSerializer
     else
       raise ExceptionHandler::ValidationError.new(@artists_collaborator.errors.to_h, 'Error updating collaborator.')
     end
@@ -35,8 +35,10 @@ class Api::V1::ArtistsCollaboratorsController < Api::BaseController
 
   param_group :doc_destroy_artists_collaborator
   def destroy
+    raise ExceptionHandler::BadRequest, 'Cannot delete collaborator' unless @artists_collaborator.collaborator.tracks.blank?
+
     if @artists_collaborator.destroy
-      render json: current_user.collaborators, each_serializer: Api::V1::CollaboratorSerializer
+      render json: current_user.collaborators_details.includes(:collaborator), each_serializer: Api::V1::CollaboratorSerializer
     else
       raise ExceptionHandler::ValidationError.new(@artists_colaborator.errors.to_h, 'Error deleting collaborator.')
     end
